cast to unsigned long to int64

+++ google/cloud/pubsub/internal/publisher_tracing_connection.cc
@@ -53,7 +53,7 @@ opentelemetry::nostd::shared_ptr<opentelemetry::trace::Span> StartPublishSpan(
                           {sc::kMessagingDestinationTemplate, "topic"},
                           {"messaging.pubsub.ordering_key", m.ordering_key()},
                           {"messaging.message.total_size_bytes",
-                           static_cast<int>(MessageSize(m))},
+                           static_cast<std::int64_t>(MessageSize(m))},
                       },
                       options);
   return span;
